requirements.txt
google-genai>=0.10.0
requests>=2.31.0
beautifulsoup4>=4.12.2
ttkbootstrap>=1.7.3